backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
python-jose
python-multipart

